started major refactoring for more fluid sketching
