--- Assets/Scripts/TouchAndHoldToSketch.cs.meta
@@ -1,18 +1,0 @@
-fileFormatVersion: 2
-guid: f9d0919807eaa1947936953676e285e4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - Camera: {instanceID: 0}
-  - SketchWorld: {instanceID: 0}
-  - SketchObjectPrefab: {instanceID: 0}
-  - defaults: {fileID: 11400000, guid: 80ac1fe1704d8434eac74bc2c69b9794, type: 2}
-  - ColorPicker: {instanceID: 0}
-  - BrushMarker: {instanceID: 0}
-  - AnchorPlanePrefab: {instanceID: 0}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
